+++ src/main/java/insane96mcp/enhancedai/modules/mobs/avoidexplosions/AvoidExplosions.java
@@ -1,13 +1,16 @@
-package insane96mcp.enhancedai.modules.base.avoidexplosions;
+package insane96mcp.enhancedai.modules.mobs.avoidexplosions;
 
+import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
-import insane96mcp.insanelib.base.config.Blacklist;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.insanelib.data.IdTagMatcher;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.item.PrimedTnt;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
@@ -15,9 +18,10 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.List;
 
-@Label(name = "Avoid Explosions", description = "Mobs will run away from exploding creepers / TNT")
-@LoadFeature(module = Modules.Ids.BASE)
+@Label(name = "Avoid Explosions", description = "Mobs will run away from exploding creepers / TNT. Use the entity type tag enhancedai:no_run_from_explosion to blacklist them")
+@LoadFeature(module = Modules.Ids.MOBS)
 public class AvoidExplosions extends Feature {
+	public static final TagKey<EntityType<?>> NO_RUN_FROM_EXPLOSION = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "no_run_from_explosion"));
 	@Config(min = 0d, max = 10d)
 	@Label(name = "Flee speed Multiplier Near", description = "Speed multiplier when the mob runs from explosions and it's within 7 blocks from him.")
 	public static Double runSpeedNear = 1.25d;
@@ -27,9 +31,6 @@ public class AvoidExplosions extends Feature {
 	@Config(min = 0d, max = 10d)
 	@Label(name = "Flee TNT", description = "Entities also flee from TnTs")
 	public static Boolean fleeTnt = false;
-	@Config(min = 0d, max = 10d)
-	@Label(name = "Entity Blacklist", description = "Entities that shouldn't be affected by this feature")
-	public static Blacklist entityBlacklist = new Blacklist(List.of(IdTagMatcher.newId("minecraft:warden")), false);
 
 	public AvoidExplosions(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
@@ -37,7 +38,8 @@ public class AvoidExplosions extends Feature {
 
 	@SubscribeEvent
 	public void onMobSpawn(EntityJoinLevelEvent event) {
-		if (!this.isEnabled())
+		if (!this.isEnabled()
+				|| event.getEntity().getType().is(NO_RUN_FROM_EXPLOSION))
 			return;
 
 		addAvoidAI(event);
@@ -45,8 +47,7 @@ public class AvoidExplosions extends Feature {
 	}
 
 	private void addAvoidAI(EntityJoinLevelEvent event) {
-		if (!(event.getEntity() instanceof PathfinderMob creatureEntity)
-				|| entityBlacklist.isEntityBlackOrNotWhitelist(creatureEntity))
+		if (!(event.getEntity() instanceof PathfinderMob creatureEntity))
 			return;
 
 		creatureEntity.goalSelector.addGoal(1, new AvoidExplosionGoal(creatureEntity, runSpeedNear, runSpeedFar));